Separate icon tests
